+++ crates/topcoat-core/src/context.rs
@@ -1,15 +1,17 @@
+mod binding;
 mod context_map;
 mod id;
 
 use std::{
     any::{Any, TypeId},
     cell::RefCell,
-    collections::{HashMap, HashSet},
+    collections::HashSet,
     marker::PhantomData,
     ops::Deref,
     sync::{Arc, Mutex},
 };
 
+use binding::{BindingMask, BindingRegistry, ContextBindingId};
 use bit_set::BitSet;
 pub use context_map::*;
 pub use id::*;
@@ -404,71 +406,6 @@ impl CxTestBuilder {
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct ContextBindingId(usize);
-
-impl ContextBindingId {
-    fn frontier(self) -> usize {
-        self.0
-            .checked_add(1)
-            .expect("request context binding ID overflowed")
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum BindingKind {
-    RootNone,
-    Value,
-}
-
-#[derive(Debug, Clone, Default)]
-struct BindingMask {
-    bits: BitSet<usize>,
-    frontier: usize,
-}
-
-impl BindingMask {
-    fn install(
-        &mut self,
-        registry: &BindingRegistry,
-        type_id: TypeId,
-        previous_id: Option<ContextBindingId>,
-    ) -> ContextBindingId {
-        let (root_none_id, binding_id) = registry.allocate_value(type_id);
-        self.advance(registry, binding_id.frontier());
-        let previous_id = previous_id.unwrap_or(root_none_id);
-        let removed = self.bits.remove(previous_id.0);
-        debug_assert!(removed, "previous context binding was not visible");
-        self.bits.insert(binding_id.0);
-        binding_id
-    }
-
-    fn advance(&mut self, registry: &BindingRegistry, frontier: usize) {
-        assert!(
-            frontier >= self.frontier,
-            "request context binding mask cannot move backwards"
-        );
-        for index in self.frontier..frontier {
-            if registry.kind(ContextBindingId(index)) == BindingKind::RootNone {
-                self.bits.insert(index);
-            }
-        }
-        self.frontier = frontier;
-    }
-
-    fn effectively_contains(
-        &self,
-        registry: &BindingRegistry,
-        binding_id: ContextBindingId,
-    ) -> bool {
-        if binding_id.0 < self.frontier {
-            self.bits.contains(binding_id.0)
-        } else {
-            registry.kind(binding_id) == BindingKind::RootNone
-        }
-    }
-}
-
 struct ContextBinding {
     id: ContextBindingId,
     value: ContextValue,
@@ -486,49 +423,6 @@ impl ContextBinding {
             .value
             .downcast::<T>()
             .expect("context binding type changed")
-    }
-}
-
-#[derive(Debug, Default)]
-struct BindingRegistry {
-    root_none: Mutex<HashMap<TypeId, ContextBindingId>>,
-    kinds: boxcar::Vec<BindingKind>,
-}
-
-impl BindingRegistry {
-    fn root_none(&self, type_id: TypeId) -> ContextBindingId {
-        let mut root_none = self.root_none.lock().unwrap();
-        *root_none
-            .entry(type_id)
-            .or_insert_with(|| self.push(BindingKind::RootNone))
-    }
-
-    fn allocate_value(&self, type_id: TypeId) -> (ContextBindingId, ContextBindingId) {
-        let mut root_none = self.root_none.lock().unwrap();
-        let root_none_id = *root_none
-            .entry(type_id)
-            .or_insert_with(|| self.push(BindingKind::RootNone));
-        (root_none_id, self.push(BindingKind::Value))
-    }
-
-    fn push(&self, kind: BindingKind) -> ContextBindingId {
-        self.kinds
-            .count()
-            .checked_add(1)
-            .expect("request context binding ID overflowed");
-        ContextBindingId(self.kinds.push(kind))
-    }
-
-    fn kind(&self, binding_id: ContextBindingId) -> BindingKind {
-        *self
-            .kinds
-            .get(binding_id.0)
-            .expect("request context binding metadata was not initialized")
-    }
-
-    #[cfg(test)]
-    fn count(&self) -> usize {
-        self.kinds.count()
     }
 }
 
@@ -646,281 +540,4 @@ pub fn memoize_cache(cx: &Cx) -> &MemoizeCache {
 #[must_use]
 pub fn abort_store(cx: &Cx) -> &AbortStore {
     &cx.request_state.abort_store
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::{
-        Barrier,
-        atomic::{AtomicUsize, Ordering as AtomicOrdering},
-    };
-
-    use super::*;
-
-    #[derive(Debug, PartialEq)]
-    struct Database(&'static str);
-
-    #[derive(Debug, PartialEq)]
-    struct Config(u32);
-
-    struct DropCounter(Arc<AtomicUsize>);
-
-    impl Drop for DropCounter {
-        fn drop(&mut self) {
-            self.0.fetch_add(1, AtomicOrdering::SeqCst);
-        }
-    }
-
-    #[test]
-    fn concurrent_missing_reads_share_one_root_none_identity() {
-        let cx = Cx::default();
-        let ready = Barrier::new(3);
-
-        let (first_id, second_id) = std::thread::scope(|scope| {
-            let first = scope.spawn(|| {
-                ready.wait();
-                assert_eq!(try_request_context::<Database>(&cx), None);
-                cx.resolve_binding_id(TypeId::of::<Database>())
-            });
-            let second = scope.spawn(|| {
-                ready.wait();
-                assert_eq!(try_request_context::<Database>(&cx), None);
-                cx.resolve_binding_id(TypeId::of::<Database>())
-            });
-            ready.wait();
-            (first.join().unwrap(), second.join().unwrap())
-        });
-
-        assert_eq!(first_id, second_id);
-        assert_eq!(cx.request_state.bindings.count(), 1);
-        assert_eq!(
-            cx.request_state.bindings.kind(first_id),
-            BindingKind::RootNone
-        );
-    }
-
-    #[test]
-    fn contexts_created_before_root_none_treat_it_as_visible() {
-        let cx = Cx::default();
-        let before = cx.with(Config(1));
-        let root_none_id = cx
-            .request_state
-            .bindings
-            .root_none(TypeId::of::<Database>());
-
-        assert!(root_none_id.0 >= before.binding_mask.frontier);
-        assert!(!before.binding_mask.bits.contains(root_none_id.0));
-        assert!(
-            before
-                .binding_mask
-                .effectively_contains(&before.request_state.bindings, root_none_id)
-        );
-        assert_eq!(try_request_context::<Database>(&before), None);
-    }
-
-    #[test]
-    fn first_value_allocates_and_shadows_root_none() {
-        let cx = Cx::default();
-        let child = cx.with(Database("primary"));
-        let root_none_id = cx
-            .request_state
-            .bindings
-            .root_none(TypeId::of::<Database>());
-        let value_id = child.resolve_binding_id(TypeId::of::<Database>());
-
-        assert_eq!(root_none_id, ContextBindingId(0));
-        assert_eq!(value_id, ContextBindingId(1));
-        assert!(
-            cx.binding_mask
-                .effectively_contains(&cx.request_state.bindings, root_none_id)
-        );
-        assert!(!child.binding_mask.bits.contains(root_none_id.0));
-        assert!(child.binding_mask.bits.contains(value_id.0));
-    }
-
-    #[test]
-    fn advancing_for_an_unrelated_value_materializes_root_none() {
-        let cx = Cx::default();
-        let root_none_id = cx
-            .request_state
-            .bindings
-            .root_none(TypeId::of::<Database>());
-
-        assert!(!cx.binding_mask.bits.contains(root_none_id.0));
-        let child = cx.with(Config(1));
-
-        assert!(child.binding_mask.bits.contains(root_none_id.0));
-        assert_eq!(try_request_context::<Database>(&child), None);
-    }
-
-    #[test]
-    fn root_insert_replaces_and_returns_the_displaced_value() {
-        let mut cx = Cx::default();
-
-        assert_eq!(cx.insert(Database("primary")), None);
-        let root_none_id = cx
-            .request_state
-            .bindings
-            .root_none(TypeId::of::<Database>());
-        let first_id = cx.resolve_binding_id(TypeId::of::<Database>());
-        assert!(!cx.binding_mask.bits.contains(root_none_id.0));
-        assert!(cx.binding_mask.bits.contains(first_id.0));
-        assert_eq!(cx.insert(Database("replica")), Some(Database("primary")));
-        let second_id = cx.resolve_binding_id(TypeId::of::<Database>());
-        assert_eq!(request_context::<Database>(&cx), &Database("replica"));
-        assert_ne!(first_id, second_id);
-        assert!(!cx.binding_mask.bits.contains(first_id.0));
-        assert!(cx.binding_mask.bits.contains(second_id.0));
-    }
-
-    #[test]
-    fn get_mut_changes_a_root_value() {
-        let mut cx = CxTestBuilder::new().request_context(Config(1)).build();
-
-        let first_id = cx.resolve_binding_id(TypeId::of::<Config>());
-        cx.get_mut::<Config>().unwrap().0 = 42;
-        let second_id = cx.resolve_binding_id(TypeId::of::<Config>());
-
-        assert_eq!(request_context::<Config>(&cx), &Config(42));
-        assert_ne!(first_id, second_id);
-        assert!(!cx.binding_mask.bits.contains(first_id.0));
-        assert!(cx.binding_mask.bits.contains(second_id.0));
-        let binding_count = cx.request_state.bindings.count();
-        let mask = cx.binding_mask.clone();
-        assert_eq!(cx.get_mut::<Database>(), None);
-        assert_eq!(cx.request_state.bindings.count(), binding_count);
-        assert_eq!(cx.binding_mask.bits, mask.bits);
-        assert_eq!(cx.binding_mask.frontier, mask.frontier);
-    }
-
-    #[test]
-    fn child_scope_shadows_parent_without_changing_it() {
-        let mut cx = Cx::default();
-        cx.insert(Database("primary"));
-        let child = cx.with(Database("replica"));
-
-        assert_eq!(request_context::<Database>(&child), &Database("replica"));
-        assert_eq!(request_context::<Database>(&cx), &Database("primary"));
-    }
-
-    #[test]
-    fn nearest_scoped_binding_wins() {
-        let cx = Cx::default();
-        let child = cx.with(Database("replica"));
-        let grandchild = child.with(Database("archive"));
-
-        assert_eq!(
-            request_context::<Database>(&grandchild),
-            &Database("archive")
-        );
-        assert_eq!(request_context::<Database>(&child), &Database("replica"));
-    }
-
-    #[test]
-    fn unrelated_values_are_inherited() {
-        let cx = CxTestBuilder::new().request_context(Config(42)).build();
-        let child = cx.with(Database("replica"));
-
-        assert_eq!(request_context::<Config>(&child), &Config(42));
-    }
-
-    #[test]
-    fn dropping_a_scope_leaves_the_parent_unchanged() {
-        let mut cx = Cx::default();
-        {
-            let child = cx.with(Database("replica"));
-            assert_eq!(request_context::<Database>(&child), &Database("replica"));
-        }
-
-        assert_eq!(try_request_context::<Database>(&cx), None);
-        cx.insert(Database("primary"));
-        assert_eq!(request_context::<Database>(&cx), &Database("primary"));
-    }
-
-    #[test]
-    fn dropping_a_scope_drops_only_its_shadowing_binding() {
-        let parent_drops = Arc::new(AtomicUsize::new(0));
-        let child_drops = Arc::new(AtomicUsize::new(0));
-        let cx = Cx::default();
-        let parent = cx.with(DropCounter(parent_drops.clone()));
-
-        {
-            let child = parent.with(DropCounter(child_drops.clone()));
-            assert!(Arc::ptr_eq(
-                &request_context::<DropCounter>(&child).0,
-                &child_drops,
-            ));
-        }
-
-        assert_eq!(child_drops.load(AtomicOrdering::SeqCst), 1);
-        assert_eq!(parent_drops.load(AtomicOrdering::SeqCst), 0);
-        drop(parent);
-        assert_eq!(parent_drops.load(AtomicOrdering::SeqCst), 1);
-    }
-
-    #[test]
-    fn with_values_installs_each_tuple_element() {
-        let cx = Cx::default();
-        let child = cx.with_values((Database("primary"), Config(42)));
-
-        assert_eq!(request_context::<Database>(&child), &Database("primary"));
-        assert_eq!(request_context::<Config>(&child), &Config(42));
-    }
-
-    #[test]
-    fn with_treats_a_tuple_as_one_binding() {
-        let cx = Cx::default();
-        let child = cx.with((Database("primary"), Config(42)));
-
-        assert_eq!(
-            request_context::<(Database, Config)>(&child),
-            &(Database("primary"), Config(42))
-        );
-        assert_eq!(try_request_context::<Database>(&child), None);
-    }
-
-    #[test]
-    #[should_panic(expected = "duplicate value types")]
-    fn with_values_rejects_duplicate_types() {
-        let cx = Cx::default();
-        let _ = cx.with_values((Config(1), Config(2)));
-    }
-
-    #[test]
-    fn with_values_supports_arities_two_through_twelve() {
-        let cx = Cx::default();
-        let _ = cx.with_values((0u8, 1u16));
-        let _ = cx.with_values((0u8, 1u16, 2u32));
-        let _ = cx.with_values((0u8, 1u16, 2u32, 3u64));
-        let _ = cx.with_values((0u8, 1u16, 2u32, 3u64, 4u128));
-        let _ = cx.with_values((0u8, 1u16, 2u32, 3u64, 4u128, 5i8));
-        let _ = cx.with_values((0u8, 1u16, 2u32, 3u64, 4u128, 5i8, 6i16));
-        let _ = cx.with_values((0u8, 1u16, 2u32, 3u64, 4u128, 5i8, 6i16, 7i32));
-        let _ = cx.with_values((0u8, 1u16, 2u32, 3u64, 4u128, 5i8, 6i16, 7i32, 8i64));
-        let _ = cx.with_values((0u8, 1u16, 2u32, 3u64, 4u128, 5i8, 6i16, 7i32, 8i64, 9i128));
-        let _ = cx.with_values((
-            0u8, 1u16, 2u32, 3u64, 4u128, 5i8, 6i16, 7i32, 8i64, 9i128, 10usize,
-        ));
-        let _ = cx.with_values((
-            0u8, 1u16, 2u32, 3u64, 4u128, 5i8, 6i16, 7i32, 8i64, 9i128, 10usize, 11isize,
-        ));
-    }
-
-    #[test]
-    fn each_scope_has_a_fresh_id() {
-        let cx = Cx::default();
-        let child = cx.with(Config(1));
-        let grandchild = child.with(Database("primary"));
-
-        assert_ne!(cx.id(), child.id());
-        assert_ne!(child.id(), grandchild.id());
-    }
-
-    #[test]
-    fn context_types_are_send_and_sync() {
-        fn assert_send_sync<T: Send + Sync>() {}
-
-        assert_send_sync::<Cx>();
-        assert_send_sync::<CxScope<'static>>();
-    }
 }
